+++ internal/pkg/dto/payment.go
@@ -8,7 +8,7 @@ type PaymentDto struct {
 	PaymentId  string
 	UserId     string
 	MerchantId string
-	Amount     int64
+	Amount     float32
 }
 
 func (p *PaymentDto) FromEntity(entity payment.Payment) PaymentDto {
